enhance: split work into pieces
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@24427 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ net/enhance.py
@@ -22,12 +22,10 @@ from astrometry.util.plotutils import *
 from astrometry.util.miscutils import *
 
 import logging
-logging.basicConfig(format='%(message)s',
-                    level=logging.DEBUG)
 
 topscale = 256.
 
-def addcal(cal, version):
+def addcal(cal, version, hpmod, hpnum):
     tan = cal.raw_tan
     try:
         if not cal.job.user_image.publicly_visible:
@@ -40,7 +38,7 @@ def addcal(cal, version):
         return False
 
     ft = cal.job.user_image.image.disk_file.file_type
-    print 'File type:', ft
+    #print 'File type:', ft
 
     # HACK
     #if not 'image' in ft:
@@ -49,11 +47,11 @@ def addcal(cal, version):
         return False
 
     wcsfn = cal.get_wcs_file()
-    print 'WCS file:', wcsfn
+    #print 'WCS file:', wcsfn
     df = cal.job.user_image.image.disk_file
-    print 'Original filename:', cal.job.user_image.original_file_name
-    print 'Submission:', cal.job.user_image.submission
-    print 'DiskFile', df
+    #print 'Original filename:', cal.job.user_image.original_file_name
+    #print 'Submission:', cal.job.user_image.submission
+    #print 'DiskFile', df
     ft = df.file_type
     fn = df.get_path()
     if 'JPEG' in ft:
@@ -66,13 +64,13 @@ def addcal(cal, version):
         u = np.unique(I.ravel())
         print 'Number of unique pixel values:', len(u)
         if I.dtype != np.uint8:
-            print 'Datatype:', I.dtype
+            #print 'Datatype:', I.dtype
             return False
         # arbitrary value!
         #if len(u) <= 25:
         #    continue
     wcs = Sip(wcsfn)
-    print 'WCS', wcs
+    #print 'WCS', wcs
 
     print 'Pixscale', tan.get_pixscale()
     nside = 2 ** int(np.round(np.log2(topscale / tan.get_pixscale())))
@@ -89,6 +87,13 @@ def addcal(cal, version):
     r,d,radius = tan.get_center_radecradius()
     radius = np.hypot(radius, hpradius)
     hh = healpix_rangesearch_radec(r, d, radius, nside)
+    hh.sort()
+    print 'Healpixes:', hh
+
+    if hpmod:
+        hh = [h for h in hh if h % hpmod == hpnum]
+        print 'Cut to healpixes:', hh
+
     for hp in hh:
         print 'Healpix', hp
         # Check for actual overlap before (possibly) creating EnhancedImage
@@ -107,13 +112,15 @@ def addcal(cal, version):
             
         if created or en.wcs is None:
             if created:
-                print 'No EnhancedImage for this nside/healpix yet'
+                # print 'No EnhancedImage for this nside/healpix yet'
+                pass
             else:
-                print 'Re-initializing', en
+                # print 'Re-initializing', en
+                pass
             en.init()
             en.save()
         else:
-            print 'EnhancedImage exists:', en
+            #print 'EnhancedImage exists:', en
             try:
                 #print 'Cals:', en.cals.all()
                 en.cals.get(id=cal.id)
@@ -134,7 +141,7 @@ def addcal(cal, version):
         K = enhM[Yo, Xo]
         Xo,Yo = Xo[K],Yo[K]
         Xi,Yi = Xi[K],Yi[K]
-        print len(Yo), 'resampled within healpix'
+        #print len(Yo), 'resampled within healpix'
         if len(Yo) == 0:
             continue
 
@@ -181,35 +188,60 @@ def addcal(cal, version):
 
 
 
-print 'DELETING ALL'
-EnhanceVersion.objects.all().delete()
-EnhancedImage.objects.all().delete()
-print 'ok'
+if __name__ == '__main__':
+    import optparse
+    parser = optparse.OptionParser('%(prog)')
+    parser.add_option('-v', dest='verbose', default=False, action='store_true')
+    parser.add_option('-D', dest='delete', action='store_true',
+                      help='Delete all database entries before beginning?')
 
-enver,created = EnhanceVersion.objects.get_or_create(name='v1', topscale=topscale)
-print 'Created', enver
+    parser.add_option('--mod', dest='mod', type=int, default=0,
+                      help='Split processing into this many pieces')
+    parser.add_option('--num', dest='num', type=int,
+                      help='Process this number out of "--mod" pieces (0-indexed)')
 
-cals = Calibration.objects.all()
-print 'Calibrations:', cals.count()
-cals = cals.select_related('raw_tan')
+    opt,args = parser.parse_args()
 
-level = 9
-slo = topscale / 2.**(level+0.5)
-shi = topscale / 2.**(level-0.5)
+    if opt.mod:
+        assert(opt.num < opt.mod)
 
-ncals = cals.count()
-for ical in range(ncals):
-    cal = cals[ical]
-    print 'Cal', cal
-    pixscale = cal.raw_tan.get_pixscale()
-    if pixscale < slo or pixscale > shi:
-        print 'Skipping: pixscale', pixscale
-        continue
+    lvl = logging.INFO
+    if opt.verbose:
+        lvl = logging.DEBUG
+    logging.basicConfig(level=lvl, format='%(message)s', stream=sys.stdout)
+    
+    if opt.delete:
+        print 'DELETING ALL'
+        EnhanceVersion.objects.all().delete()
+        EnhancedImage.objects.all().delete()
+        print 'ok'
 
-    addcal(cal, enver)
+    enver,created = EnhanceVersion.objects.get_or_create(name='v1', topscale=topscale)
+    print 'Version:', enver
 
-sys.exit(0)
+    cals = Calibration.objects.all()
+    print 'Calibrations:', cals.count()
+    cals = cals.select_related('raw_tan')
 
+    # level = 9
+    # slo = topscale / 2.**(level+0.5)
+    # shi = topscale / 2.**(level-0.5)
+
+    ncals = cals.count()
+    for ical in range(ncals):
+        print
+        print 'Calibration', ical, 'of', ncals
+        cal = cals[ical]
+        print 'Cal', cal
+
+        # pixscale = cal.raw_tan.get_pixscale()
+        # if pixscale < slo or pixscale > shi:
+        #     print 'Skipping: pixscale', pixscale
+        #     continue
+
+        addcal(cal, enver, opt.mod, opt.num)
+
+    sys.exit(0)
 
 '''
         plt.clf()
